chore: refactor gatsby-plugin-canonical-urls into a local plugin
