+++ gatsby-ssr.js
@@ -1,6 +1,3 @@
-import React from "react";
-import url from "url";
-
 /**
  * Implement Gatsby's SSR (Server Side Rendering) APIs in this file.
  *
@@ -9,25 +6,3 @@ import url from "url";
 
 // You can delete this file if you're not using it
 
-const SITE_URL = 'https://developers.cloudflare.com';
-
-exports.onRenderBody = (
-  { setHeadComponents, pathname = `/` },
-) => {
-  const siteUrl = SITE_URL.replace(/\/$/, ``)
-  const parsed = url.parse(`${siteUrl}${pathname}`)
-
-  parsed.pathname = !parsed.pathname.endsWith('/') ? '/' : parsed.pathname;
-
-  const pageUrl = `${parsed.protocol}//${parsed.host}${parsed.pathname}`;
-
-  setHeadComponents([
-    <link
-      rel="canonical"
-      key={pageUrl}
-      href={pageUrl}
-      data-baseprotocol={parsed.protocol}
-      data-basehost={parsed.host}
-    />,
-  ])
-}
