+++ plugins/gatsby-plugin-canonical-urls/gatsby-browser.js
@@ -1,0 +1,14 @@
+exports.onRouteUpdate = ({ location }) => {
+  const domElem = document.querySelector(`link[rel='canonical']`)
+  const existingValue = domElem.getAttribute(`href`)
+  const baseProtocol = domElem.getAttribute(`data-baseProtocol`)
+  const baseHost = domElem.getAttribute(`data-baseHost`)
+  
+  location.pathname = !location.pathname.endsWith('/') ? '/' : location.pathname;
+
+  let value = `${baseProtocol}//${baseHost}${location.pathname}${location.hash}`
+  if (existingValue && baseProtocol && baseHost) {
+
+    domElem.setAttribute(`href`, `${value}`)
+  }
+}
