+++ plugins/gatsby-plugin-canonical-urls/package.json
@@ -1,0 +1,3 @@
+{
+  "name": "gatsby-plugin-canonical-urls"
+}
